requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
dj-database-url==0.5.0
Django==3.1
django-bootstrap4==2.2.0
django-stdimage==5.1.1
gunicorn==20.0.4
mysql==0.0.2
mysqlclient==2.0.1
Pillow==7.2.0
progressbar2==3.51.4
psycopg2-binary==2.8.5
PyMySQL==0.10.0
python-utils==2.4.0
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
whitenoise==5.2.0
